+++ cmd/agentiik-api/start_test.go
@@ -34,6 +34,7 @@ import (
 	"github.com/agentiik/agentiik/internal/bustest"
 	"github.com/agentiik/agentiik/internal/config"
 	"github.com/agentiik/agentiik/internal/dbtest/dbname"
+	"github.com/agentiik/agentiik/internal/stopsignal"
 	"github.com/agentiik/agentiik/secret"
 	"github.com/agentiik/agentiik/version"
 	"github.com/jackc/pgx/v5"
@@ -666,7 +667,7 @@ func TestMain(m *testing.M) {
 		os.Exit(instance(raw))
 	}
 	if os.Getenv(slowStopVariable) != "" {
-		ctx, _ := signalled()
+		ctx, _ := stopsignal.Context()
 		fmt.Println("waiting")
 		<-ctx.Done()
 		fmt.Println("stopping")